Decouple OrderingScheme from AST
