+++ core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/test/PlanBuilder.java
@@ -43,6 +43,7 @@ import io.trino.sql.PlannerContext;
 import io.trino.sql.analyzer.TypeSignatureProvider;
 import io.trino.sql.parser.SqlParser;
 import io.trino.sql.planner.OrderingScheme;
+import io.trino.sql.planner.OrderingTranslator;
 import io.trino.sql.planner.Partitioning;
 import io.trino.sql.planner.PartitioningScheme;
 import io.trino.sql.planner.PlanNodeIdAllocator;
@@ -451,7 +452,7 @@ public class PlanBuilder
                     aggregation.getArguments(),
                     aggregation.isDistinct(),
                     aggregation.getFilter().map(Symbol::from),
-                    aggregation.getOrderBy().map(OrderingScheme::fromOrderBy),
+                    aggregation.getOrderBy().map(OrderingTranslator::fromOrderBy),
                     mask));
         }
 
@@ -1300,7 +1301,7 @@ public class PlanBuilder
                 aggregation.getArguments(),
                 aggregation.isDistinct(),
                 aggregation.getFilter().map(Symbol::from),
-                aggregation.getOrderBy().map(OrderingScheme::fromOrderBy),
+                aggregation.getOrderBy().map(OrderingTranslator::fromOrderBy),
                 Optional.empty());
     }
